Cygwin: support custom package names
Different package managers package Camlp4 under different names. For
example, Cygwin and MacPorts use ocaml-camlp4, but Homebrew uses
camlp4.

I am not sure if it is better to always use the underlying package
manager's name, or to give a uniform name in the ocaml-binaries
interface. I am settling on a uniform name, as it may simplify the
build matrix. Choosing camlp4.